API

PlotMGA.EnergyUnit — Type

EnergyUnit MWh=1 GWh=1000

The multiplicities of SpineOpt's native flow unit - MWh

PlotMGA.ProfileType — Type

struct ProfileType
    name::String # profile name
    df_predicate::Pair # df filter predicate
end

Describes a single flow profile. The predicate greps all of the flows from a master flow df.

Example

type1 = ProfileType("Electricity", :destination=>ByRow(contains("Electricity")))

type2 = ProfileType("Heat", :destination=>ByRow(contains("Space_Heating")))

df_electricity = subset(master_df, type1.df_predicate)

df_heat = subset(master_df, type2.df_predicate)

PlotMGA.add_mga_iteration — Method

add_mga_iteration(params::DataFrame)

Adds mga_iteration with the MGA iteration number to a df with details of a SpineOpt's entity.

The SpineOpt's MGA data should always the last entry in details, with a format "mgaitNUMBER".

PlotMGA.annotate_df_by_profiles — Method

annotate_df_by_profiles(df, combined_profiles::Vector{CombinedProfileTypes})

Adds a column type based on fulfilling the combined profile predicates.

The entries not fulfilling any predicates will be discarded.

Entries fulfilling multiple combinations will be multiplied.

PlotMGA.annotate_optimal_solutions — Method

annotate_optimal_solutions(df)

Changes the alternative field for optimal solutions from the MGA process.

PlotMGA.by_pairs — Method

by_pairs(df, group_cols)

Groups df and create the iterator in the form group_column_values, subdf

PlotMGA.change_col_units — Method

change_col_units(df, unit::EnergyUnit, column::Symbol)

Changes the colums energy units to those passed.

PlotMGA.combine_profiles — Method

combine_profiles(profile_types::Vector{ProfileType}...)

Produces all combinations of flow profile types from multiple lists.

PlotMGA.connect_technology_with_effects — Method

connect_technology_with_effects(df, effects_df)

Assigns total system effects to iterations from appropriate alternatives.

For example, connects technology installation with total system costs.

PlotMGA.get_spine_parameter — Method

get_spine_parameter(db, param_name::Symbol[, col_order])

Grabs the parameter from a SpineOpt's database and processess it according to the parameter name.

The option parameter col_order specifies the only columns that should be in the output and their order. Those columns should be sortable.

PlotMGA.get_spine_parameter_value — Method

get_spine_parameter_value(db, param_name)

Get parameter value from a db in a SpineDB format.

Returns a df with columns:

alternative - SpineDB alternative (e.g., used method)

value - parameter value

details - list of fields from the Spine's entity name (e.g., parameter name, mga iteration, technology)

PlotMGA.obtain_technology_interactions — Method

obtain_technology_interactions(df, effects_df)

Produces a df describing a matrix of interactions between different technology installations and their effects (e.g., total system costs).

PlotMGA.plot_all_mga_profiles — Method

plot_all_mga_profiles(
    df, unit::EnergyUnit, profile_types::Vector{ProfileType}...;
    fig_prefix::String="", display_plot::Bool=true, save_plot::Bool=true
)

Plots all profiles for every flow type combination and every alternative from the df column alternative.

PlotMGA.plot_mga_alternative_profiles — Method

plot_mga_alternative_profiles(
    df, unit::EnergyUnit;
    fig_prefix::String="", display_plot::Bool=true, save_plot::Bool=true
)

Plots flow profiles of a single type across all of the alternatives present in df field alternative. Every alternative produces a single figure.

If the figures are to be saved fig_prefix is added to the filename.

PlotMGA.plot_mga_profile — Method

plot_mga_profile(
    df, unit::EnergyUnit;
    iteration_col::Symbol=:mga_iteration, value_col::Symbol=:total_value, stackby::Symbol=:technology
)

Plots stacked flow profiles for a df describing a single MGA run.

PlotMGA.plot_technology_effects — Method

plot_technology_effects(
    tech_df, effects_df;
    fig_prefix::String="", display_plot::Bool=true, save_plot::Bool=true
)

Creates a plot for every technology and show effects of its installation on the total system state (e.g., total system_costs)

If the figures are to be saved fig_prefix is added to the filename.

PlotMGA.plot_technology_effects_by_alternative — Method

plot_technology_effects_by_alternative(
    df, plot;
    technology_col::Symbol=:total_value, effect_col::Symbol=:total_value_effects
)

Modifies the passed plot by adding point corresponding to the technology and its effects in every alternative.

PlotMGA.plot_technology_interactions — Method

plot_technology_interactions(
    tech_df, effects_df, alternative_markers::AbstractDict;
    fig_prefix::String="", display_plot::Bool=true, save_plot::Bool=true
)

Creates a plot for every pair of the technologies, showing interactions between their installed capacities and effects (e.g., total system costs). Every plot has entries from all of the alternatives from df.

To tell apart the alternatives, we have to pass Dict of markers.

If the figures are to be saved fig_prefix is added to the filename.

PlotMGA.plot_technology_interactions_by_alternative — Method

plot_technology_interactions_by_alternative(
    df, plot, alternative_markers::AbstractDict;
    x_technology_col::Symbol=:total_value_1st, y_technology_col::Symbol=:total_value_2nd, effect_col::Symbol=:total_value_effects
)

Modifies the passed plot by adding point corresponding to the technology interactions per every alternative. The points are colored according to the effect_col from df.

To tell apart values from different alternatives (e.g., used algorithm) we have to pass dict of markers with an entry for each alternative.

PlotMGA.process_flows — Method

process_flows(params::DataFrame)

Adds `destination` to a df with the `details` of a SpineOpt's entity, and keeps only inflows.

The MORPHE2US destination data should always comes as a third in `details`.

PlotMGA.process_spine_data — Method

process_spine_data(::Val{DB_ENTRY}) where {DB_ENTRY}

Returns processing pipeline for the database entry.

It currently has a specific implementation for:
    - `total_costs`
    - `storages_invested`
    - `units_invested`
    - `unit_flow`

But you can easily extend it on the go by defining:
```julia
function process_fn1(df)
    ...
    df
end
function process_fn2(df)
    ...
    df
end
```
process_spine_data(::Val{:new_type}) = process_fn1 ∘ process_fn2

Some of the predefined functions for processing include: 
    - `sum_timeseries`
    - `add_mga_iteration`
    - `process_technology`
    - `process_flows`

PlotMGA.process_technology — Method

process_technology(params::DataFrame)

Adds technology and installation to a df with the details of a SpineOpt's entity.

The MORPHE2US technology data should always comes as a second in details, with a format "TECHNOLOGY_NAME(B|D)LVL`INSTALLATIONSITE`".

PlotMGA.squash_groups — Method

squash_groups(df, kept_cols[, combine_col])

Keep only columns with unique values of kept_cols, and squash the rest.

If combine_cols is passed, the column values are summed. If not, only the first entry is retained.

PlotMGA.sum_timeseries — Method

sum_timeseries(params::DataFrame)

Adds total_value with sum of values to a df with value.

Each of the values should be a Dict with datetime as keys and floats as values.